requirements.txt
beautifulsoup4==4.6.3
certifi==2018.8.24
future==0.16.0
python-telegram-bot==10.1.0
